+++ src/Controller/CodificationController.php
@@ -22,10 +22,15 @@ class CodificationController extends AbstractController
 	
 	public function index(CodificationRepository $codificationRepository, Project $project): Response
 	{
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		return $this->render('generic/list.html.twig', [
 			'header' => $this->translator->trans('Codification for') . ' : ' . $project->getName(),
 			'route_back' =>  $this->generateUrl('project_view', [
-				'id' => $project->getId(),
+				'project' => $project->getId(),
 			]),
 			'class' => Codification::class,
 			'entities' => $codificationRepository->getCodifications($project),
@@ -34,6 +39,11 @@ class CodificationController extends AbstractController
 
 	public function new(Request $request, Project $project): Response
 	{
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		$codification = new Codification();
 		$codification->setProject($project);
 		
@@ -47,13 +57,13 @@ class CodificationController extends AbstractController
 
 			$this->addFlash('success', 'New entry created');
 			return $this->redirectToRoute('codification', [
-				'id' => $project->getId()
+				'project' => $project->getId()
 			]);
 		} else {
 			$view = $form->createView();
 			return $this->render('generic/form.html.twig', [
 				'route_back' =>  $this->generateUrl('codification', [
-					'id' => $project->getId(),
+					'project' => $project->getId(),
 				]),
 				'form' => $view
 			]);
@@ -62,6 +72,12 @@ class CodificationController extends AbstractController
 
 	public function edit(Request $request, Codification $codification): Response
 	{
+		$project = $codification->getProject();
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		$form = $this->createForm(CodificationType::class, $codification);
 		$form->handleRequest($request);
 
@@ -70,13 +86,13 @@ class CodificationController extends AbstractController
 			$entityManager->flush();
 			$this->addFlash('success', 'Datas updated');
 			return $this->redirectToRoute('codification', [
-				'id' => $codification->getProject()->getId()
+				'project' => $codification->getProject()->getId()
 			]);
 		} else {
 			$view = $form->createView();
 			return $this->render('generic/form.html.twig', [
 				'route_back' =>  $this->generateUrl('codification', [
-					'id' => $codification->getProject()->getId(),
+					'project' => $codification->getProject()->getId(),
 				]),
 				'form' => $view
 			]);
@@ -85,6 +101,12 @@ class CodificationController extends AbstractController
 
 	public function delete(Request $request, Codification $codification): Response
 	{
+		$project = $codification->getProject();
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		if ($this->isCsrfTokenValid('delete', $request->request->get('_token'))) {
 			$entityManager = $this->getDoctrine()->getManager();
 			$entityManager->remove($codification);
@@ -92,12 +114,12 @@ class CodificationController extends AbstractController
 
 			$this->addFlash('success', 'Entry deleted');
 			return $this->redirectToRoute('codification', [
-				'id' => $codification->getProject()->getId()
+				'project' => $codification->getProject()->getId()
 			]);
 		} else {
 			return $this->render('generic/delete.html.twig', [
 				'route_back' =>  $this->generateUrl('codification', [
-					'id' => $codification->getProject()->getId(),
+					'project' => $codification->getProject()->getId(),
 						]),
 			    'entities' => [$codification],
             ]);
