+++ src/Controller/MetadataItemController.php
@@ -5,6 +5,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Contracts\Translation\TranslatorInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use App\Entity\MetadataItem;
 use App\Form\MetadataItemType;
 use App\Repository\MetadataItemRepository;
@@ -22,10 +23,16 @@ class MetadataItemController extends AbstractController
 	
 	public function index(MetadataItemRepository $metadataItemRepository, Metadata $metadata): Response
 	{
+		$project = $metadata->getProject();
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		return $this->render('generic/list.html.twig', [
 			'header' => $this->translator->trans('List for the metadata') . ' : ' . $metadata->getName(),
 			'route_back' =>  $this->generateUrl('metadata', [
-				'id' => $metadata->getProject()->getId(),
+				'project' => $metadata->getProject()->getId(),
 			]),
 			'class' => MetadataItem::class,
 			'entities' => $metadataItemRepository->getMetadataItem($metadata),
@@ -34,6 +41,12 @@ class MetadataItemController extends AbstractController
 
 	public function new(Request $request, Metadata $metadata): Response
 	{
+		$project = $metadata->getProject();
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		$metadataItem = new MetadataItem();
 		$metadataItem->setMetadata($metadata);
 		$form = $this->createForm(MetadataItemType::class, $metadataItem);
@@ -46,21 +59,30 @@ class MetadataItemController extends AbstractController
 
 			$this->addFlash('success', 'New entry created');
 			return $this->redirectToRoute('metadata_item', [
-				'id' => $metadata->getId()
+				'metadata' => $metadata->getId()
 			]);
 		} else {
 			$view = $form->createView();
 			return $this->render('generic/form.html.twig', [
 				'route_back' =>  $this->generateUrl('metadata_item', [
-					'id' => $metadata->getId(),
+					'metadata' => $metadata->getId(),
 				]),
 				'form' => $view
 			]);
 		}
 	}
-
+	
+	/**
+	 * @ParamConverter("metadataItem", options={"mapping": {"metadata_item" : "id"}})
+	 */
 	public function edit(Request $request, MetadataItem $metadataItem): Response
 	{
+		$project = $metadataItem->getMetadata()->getProject();
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		$form = $this->createForm(MetadataItemType::class, $metadataItem);
 		$form->handleRequest($request);
 
@@ -70,21 +92,30 @@ class MetadataItemController extends AbstractController
 
 			$this->addFlash('success', 'Datas updated');
 			return $this->redirectToRoute('metadata_item', [
-				'id' => $metadataItem->getMetadata()->getId()
+				'metadata' => $metadataItem->getMetadata()->getId()
 			]);
 		} else {
 			$view = $form->createView();
 			return $this->render('generic/form.html.twig', [
 				'route_back' =>  $this->generateUrl('metadata_item', [
-					'id' => $metadataItem->getMetadata()->getId(),
+					'metadata' => $metadataItem->getMetadata()->getId(),
 				]),
 				'form' => $view
 			]);
 		}
 	}
-
+	
+	/**
+	 * @ParamConverter("metadataItem", options={"mapping": {"metadata_item" : "id"}})
+	 */
 	public function delete(Request $request, MetadataItem $metadataItem): Response
 	{
+		$project = $metadataItem->getMetadata()->getProject();
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		if ($this->isCsrfTokenValid('delete', $request->request->get('_token'))) {
 			$entityManager = $this->getDoctrine()->getManager();
 			$entityManager->remove($metadataItem);
@@ -92,12 +123,12 @@ class MetadataItemController extends AbstractController
 
 			$this->addFlash('success', 'Entry deleted');
 			return $this->redirectToRoute('metadata_item', [
-				'id' => $metadataItem->getMetadata()->getId()
+				'metadata' => $metadataItem->getMetadata()->getId()
 			]);
 		} else {
 			return $this->render('generic/delete.html.twig', [
 				'route_back' =>  $this->generateUrl('metadata_item', [
-					'id' => $metadataItem->getMetadata()->getId(),
+					'metadata' => $metadataItem->getMetadata()->getId(),
 				]),
 				'entities' => [$metadataItem],
 			]);
