+++ src/Controller/ProjectController.php
@@ -34,10 +34,21 @@ class ProjectController extends AbstractController
 	public function index(): Response
 	{
 		if ($this->isGranted('ROLE_ADMIN')) {
+			
 			$projects = $this->projectRepository->getAllProjects();
+			
 		} else {
+			
 			$projects = $this->projectRepository->getProjects($this->getUser());
+			
+			if (sizeof($projects) == 1) {
+				return $this->redirectToRoute('project_view', [
+					'project' => $projects[0]->getId(),
+				]);
+			}
+			
 		}
+		
 		return $this->render('project/index.html.twig', [
 			'projects' => $projects
 		]);
@@ -45,13 +56,44 @@ class ProjectController extends AbstractController
 
 	public function view(Request $request, Project $project, CompanyRepository $companyRepository): Response
 	{
-		if ($this->getUser()->getCompany()->getType() == Company::SUPPLIER) {
+		if ($this->isGranted('ROLE_ADMIN') === false && $project->hasUser($this->getUser()) === false) {
 			return $this->redirectToRoute('project');
 		}
-		$mainContractors = $companyRepository->getMainContractors($project);
-		$subContractors= $companyRepository->getSubContractors($project);
-		$automations = $this->automationRepository->getEnabledAutomations($project);
-	
+		
+		if ($this->isGranted('ROLE_ADMIN') ||
+			$this->isGranted('ROLE_CONTROLLER') && $this->getUser()->getCompany()->isMainContractor() ||
+			$this->isGranted('ROLE_EDIT_DOCUMENTS') && $project->hasUser($this->getUser())) {
+			
+			$mainContractors = $companyRepository->getMainContractors($project);
+			$subContractors= $companyRepository->getSubContractors($project);
+			
+		} else {
+			
+			$mainContractors = $companyRepository->getMainContractors($project, $this->getUser());
+			$subContractors= $companyRepository->getSubContractors($project, $this->getUser());
+			
+		}
+		
+		$automations = [];
+		if ($this->isGranted('ROLE_ADMIN') ||
+			$this->isGranted('ROLE_CONTROLLER') && $this->getUser()->getCompany()->isMainContractor()) {
+				$automations = $this->automationRepository->getEnabledAutomations($project);
+		}
+		
+		if ($this->isGranted('ROLE_ADMIN') === false) {
+			$projects = $this->projectRepository->getProjects($this->getUser());
+			
+			if (sizeof($projects) == 1) {
+				return $this->render('project/view.html.twig', [
+					'project' => $project,
+					'main_contractors' => $mainContractors,
+					'sub_contractors' => $subContractors,
+					'automations' => $automations,
+					'route_back' => $this->generateUrl('home'),
+				]);
+			}
+		}
+			
 		return $this->render('project/view.html.twig', [
 			'project' => $project,
 			'main_contractors' => $mainContractors,
@@ -63,6 +105,10 @@ class ProjectController extends AbstractController
 
 	public function new(Request $request): Response
 	{
+		if ($this->isGranted('ROLE_ADMIN') === false) {
+			return $this->redirectToRoute('project');
+		}
+		
 		$project = new Project();
 		$form = $this->createForm(ProjectType::class, $project);
 		$form->handleRequest($request);
@@ -73,9 +119,7 @@ class ProjectController extends AbstractController
 			$entityManager->flush();
 
 			$this->addFlash('success', 'New entry created');
-			return $this->redirectToRoute('project_view', [
-				'id' => $project->getId()
-			]);
+			return $this->redirectToRoute('project');
 		} else {
 			$view = $form->createView();
 			return $this->render('generic/form.html.twig', [
@@ -87,6 +131,11 @@ class ProjectController extends AbstractController
 
 	public function edit(Request $request, Project $project): Response
 	{
+		if ($this->isGranted('ROLE_ADMIN') === false &&
+			($this->isGranted('ROLE_CONTROLLER') === false || $project->hasUser($this->getUser()) === false)) {
+			return $this->redirectToRoute('project');
+		}
+		
 		$form = $this->createForm(ProjectType::class, $project);
 		$form->handleRequest($request);
 		
@@ -95,15 +144,11 @@ class ProjectController extends AbstractController
 			$entityManager->flush();
 
 			$this->addFlash('success', 'Datas updated');
-			return $this->redirectToRoute('project_view', [
-				'id' => $project->getId()
-			]);
+			return $this->redirectToRoute('project');
 		} else {
 			$view = $form->createView();
 			return $this->render('generic/form.html.twig', [
-				'route_back' =>  $this->generateUrl('project_view', [
-					'id' => $project->getId()
-				]),
+				'route_back' =>  $this->generateUrl('project'),
 				'form' => $view
 			]);
 		}
@@ -111,20 +156,20 @@ class ProjectController extends AbstractController
 
 	public function delete(Request $request, Project $project): Response
 	{
+		if ($this->isGranted('ROLE_ADMIN') === false) {
+			return $this->redirectToRoute('project');
+		}
+		
 		if ($this->isCsrfTokenValid('delete', $request->request->get('_token'))) {
 			$entityManager = $this->getDoctrine()->getManager();
 			$entityManager->remove($project);
 			$entityManager->flush();
 
 			$this->addFlash('success', 'Entry deleted');
-			return $this->redirectToRoute('project_view', [
-				'id' => $project->getId()
-			]);
+			return $this->redirectToRoute('project');
         } else {
             return $this->render('generic/delete.html.twig', [
-            	'route_back' =>  $this->generateUrl('project_view', [
-            		'id' => $project->getId()
-            	]),
+            	'route_back' =>  $this->generateUrl('project'),
                 'entities' => [$project],
             ]);
         }
